+++ games/chapter4/examples/TankIcon.png
