+++ games/chapter4/examples/completetank.png
