+++ games/chapter4/solutions/hockey.py
